requirements.txt
Flask==2.0.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.0.0
Flask-WTF==0.15.0
flask-bootstrap==3.3.7
email_validator==1.1.2
ForgeryPy3